Update tests and code so that prek ts-compile-lint-ui check passes
