+++ airflow-core/src/airflow/ui/src/pages/Dag/Anomalies/Anomalies.test-msw-hooks.ts
@@ -1,0 +1,36 @@
+/*!
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+import { afterAll, afterEach, beforeAll } from "vitest";
+
+import { initAnomaliesTestI18n, server } from "./Anomalies.test-setup";
+
+let hooksRegistered = false;
+
+export const registerAnomaliesMswServerHooks = function registerAnomaliesMswServerHooks(): void {
+  if (hooksRegistered) {
+    return;
+  }
+  hooksRegistered = true;
+  beforeAll(async () => {
+    server.listen({ onUnhandledRequest: "bypass" });
+    await initAnomaliesTestI18n();
+  });
+  afterEach(() => server.resetHandlers());
+  afterAll(() => server.close());
+};
